fix: avoid circular dependency for PropController
This was causing a test to fail:
Class extends value undefined is not a constructor or null
