+++ packages/runtime/src/state/actions.ts
@@ -8,7 +8,8 @@ import { ComponentMeta } from './modules/components-meta'
 import { PropControllerDescriptor } from '../prop-controllers'
 import type { Size } from './react-builder-preview'
 import type { PropControllersHandle } from './modules/prop-controller-handles'
-import type { PropController, PropControllerMessage } from '../prop-controllers/instances'
+import type { PropControllerMessage } from '../prop-controllers/instances'
+import type { PropController } from '../prop-controllers/base'
 import type { APIResource, APIResourceType } from '../api/graphql/types'
 import type { SerializedControl } from '../builder'
 import { ElementImperativeHandle } from '../runtimes/react/element-imperative-handle'
